Build Ruby 1.8 and 1.9 via ruby-build
Ubuntu doesn't provide packages for these anymore so we need to build
them ourselves.

An older version of openssl is installed so that Ruby1.9 can be linked
to it (the OS provided version is incompatible).

Currently, Ruby1.8 is not linked to openssl.  There is some issue with
using the same openssl as Ruby1.9 (or the same mechanism to link).  A
lack of openssl support for Ruby1.8 currently doesn't seem to be an
issue.