Add jammy box using official ubuntu vagrantbox
bento/ubuntu-22.04 uses `eth0` naming for network interfaces.  This is
different than will be used for real machines. But perhaps correct for
VMs?

Not sure what our delivery mechanism for MIA looks like, but let's have
both an `eth0` and an `enp0s3` version of jammy to develop/test against.