Add playbook to use non-predictable network names
Lots places in the various code bases assume there is an `eth0` and an
`eth1`.

This playbook assumes that the virtual machine has a single network
interface, it replaces the `cloud-init` `netplan` config with a similar
one but changes the network name to `eth0`.

This is sufficient to meet some of the assumptions of the various code
bases, but is unlikely to be sufficient moving forward.