Complete WIP playbook to automate building deps
MIA has a number of dependencies that need to be custom built on a
machine similar to the MIA machine itself, such as `openssl`, `ruby` and
some number crunchers.

This commit adds a not-yet functional playbook to automate those
builds.  Eventually it will:

1. work
2. download needed source codes
3. apply needed patches
4. build the dependencies
5. upload the built dependencies to S3.

For now it serves as a place to store patches and as an aid memoir for
when I run the various commands manually.