Fix restarting services via systemctl
A systemd service can be `active/running`, `active/exited` or `failed`.
As far as monit is concerned, we are interested in (1) starting services
that are not running; and (2) restarting services that are running, but
need restarting, perhaps they are consuming too many resources or appear
hung.

Unfortunately, the systemctl commands `start`, `restart` and
`try-restart` don't quite work for these use cases.

* `start` will not start an `active/exited` service. systemd considers it to
  have already been started.
* `restart` will not restart a `failed` service.  systemd will only restart
  services which are `active`.
* `try-restart` will not restart a `failed` service.  systemd will only
  restart services which are `active`.

The monit files call a bash script to workaround these issues.