Fix fix for restarting via systemctl
The previous commit was just wrong.  The `systemctl` `restart` command
behaves in a manner consistent with what `monit start` and `monit
restart` commands expect.

That is, the service should be running after `monit start`/`monit
restart` regardless of whether it was running previously.  If the
service was running previously, it should be stopped and restarted.