Fix issue checking for DRb status via mia
`monit` uses a `check_drb_connection.sh` script to check the DRb status
of `emma`, `hacor`, `sas` and `meca`.  This script makes a HTTP request
to `mia` and `mia` checks if it can access a model for the specified
service.

Previously, everything worked as long as the expected models had been
correctly exported via romance/drb.  However, if they had not been
exported an exception would be raised when evaluating the
`app/controllers/monit_controller.rb` file.  If that exception occurred
the response would be a `500` error which the `check_drb_connection.sh`
script would interpret as "an unknown result, so best to do nothing".

Now, we (1) delay looking up the remote constants until a HTTP request
is received; and (2) return a `404` error if the constant has not been
published.  The `check_drb_connection.sh` script interprets a `404` as
"the model is missing, the service (e.g., meca) should be restarted.