Disable maggie reporting conflicts
The different user, inferred, mapped, derived and power metrics can
conflict.  That is more than one set of rules could be trying to write
to the same metric on the same device/group.

Maggie is supposed to detect this and report it.  We have a controller
action that is currently not responding to such reporting and
potentially(?) contributing to some maggie instability.

The reporting is disabled for now.