Install gems from a gem repo
Ideally this would be a static gem repo on S3.  However, the version of
Rubygems that we are using does not support installing from `s3://` URLs.

Possible solutions to this involve:

* Updating to a newer version of Rubygems.  This is likly to require updating
  to a new version of Ruby too.
* Access the S3 bucket over HTTP(S) embedding the AWS credentials in the URL.
* Making the gems stored on S3 public to avoid the credentials issue above.
* Download the gem repository from S3 and install from there.

This commit takes a simpler solution of including the gem repo in this git
repo.  This is not ideal but is a step towards a better solution.