Provide secrets via env_file
Previously, the secrets were being provided using the docker secrets
mechanism.  However, this had some issues.  In order to have the
postgres container be able to read the password secret the file on the
host machine needed to be readable by the same UID as the postgres user
in the postgres container.  There is no mechanism to change the
owner/group/mode for a secret (at least not if it is provided via a bind
mount).

The result of this was that the postgres password was exposed to
user/group 999 on the host machine.  Whoever that might be.

Now we provide secrets via an env_file.  The secrets env file on the
host is readable only by root.  However, the contents of that file are
available in the output of the `docker inspect` command.  This exposes
the secret to anyone who is a member of the `docker` group.  But it was
already available to them as they could always `docker exec` into a
running container and read the secret from there.