Better handle rails credentials
Rails master key and credentials.yml.enc are now per-site.  They also
persist redeployments of the visualisation app. The master key is stored
along with all of our other secrets.  The encrypted credentials are
stored on the docker etc volume.  Some hoops have to be jumped through
to get rails to use the path on the docker volume for both reading and
editing the credentials.

Fixes #62