Fix issue with destroying encryption keys on rebuild
If the Rails credentials.yml.enc file exists, we don't recreate it. This
preserves the encryption keys, but prevents us from updating it with
additional credentials in the future.

If we need to do the latter, perhaps it is as simple as copying the file
to the default Rails location running the commands to create the new
credentials and copying it back to the persistent location.  This
however is untested.