Install cluster builder
The example cluster type definitions are copied to
`/opt/concertim/usr/share/cluster-builder/cluster-types-available/` and
symlinks for all of them are created in
`/opt/concertim/usr/share/cluster-builder/cluster-types-enabled/`.  This
directory is then mounted on the cluster builder container.

The cluster builder container exposes its port on the host machine to
workaround an issue in how visualisation app is currently configured.

Visualisation app expects to be able to reach both user handler and
cluster builder on the same IP address, but at different ports.  The
docker compose file currently uses a bridged network and expects
services to communicate with each other via service name or alias. These
are not compatible.

The workaround is to have the cluster builder service (and later the
user handler service) expose their ports on the host.  Visualisation app
can then be configured with the hosts port on the docker bridged
network.

A better solution would be to configure visualisation app with the full
URL of the services.  Thereby allowing user handler and cluster builder
to be reached on different IPs.