Support host and port configuration for services
It looks like we need to run the containers using the equivalent of
`--network=host`.  This means we need to be able to configure which
ports the services run on.  Previously, we used a bridged network to
hide that and configured which ports were bound to the host with the
equivalent of `--port`.

The concertim services and cluster builder now support configurable host
and port.  The concertim openstack services will likely do so in the
future too.