Use postgres database
As concertim already installs a database it would be nice to have
killbill (optionally?) make use of that database.  There are essentially
two steps to doing that:

1. Convince myself that I know how to configure a posgresql database for
   killbill by switching the database image from `killbill/mariadb` to
   `killbill/postgresql`.
2. Switch from using a `db` service specified in the
   `docker-compose.yml` to using the external docker container
   `concertim-db-1`.

This commit implements (1).

However, investigation into (2) shows that that is probably more effort
than I'm willing to undertake at the moment.  The `killbill/postgresql`
image contains configuration that would need to be incorporated into
`concertim-db-1`.

Whilst that seems achievable we would also need to consider how to keep
this up-to-date as newer versions of killbill and kaui are released.
There may also be other production concerns.

I'm going to abort this work and revert this commit.  We should
reconsider this as we get closer to a `1.0` release, at which time we
should also consider what databases should our installation of killbill
support: (1) included mariadb; (2) included postgresql; (3) concertim's
postgresql; (4) some other external mariadb; (5) some other external
postgresql?  Also should support for an external postgresql database be
added to concertim?