Add user_handler extra host to cluster builder
The default configuration for concertim assumes that the middleware
services are available on a host called `user_handler`.  This is made
true by docker extra hosts setting.

Concertim now passes the URL for the user handler to cluster builder, so
we need to have its container configured with the extra hosts too.